requirements.txt
python-dotenv
gunicorn
flask
redis